+++ runtime/service/src/main/java/org/apache/polaris/service/events/PolarisEventListenerConfiguration.java
@@ -20,6 +20,8 @@ package org.apache.polaris.service.events;
 
 import io.quarkus.runtime.annotations.StaticInitSafe;
 import io.smallrye.config.ConfigMapping;
+import io.smallrye.config.WithParentName;
+import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import org.apache.polaris.service.events.listeners.PolarisEventListener;
@@ -31,14 +33,38 @@ public interface PolarisEventListenerConfiguration {
    * The type of the event listener to use. Must be a registered {@link PolarisEventListener}
    * identifier.
    *
-   * @deprecated since 1.5.0, use 'polaris.event-listener.types' instead
+   * @deprecated since 1.5.0, use 'polaris.event-listener.types' instead, if both are set, then
+   *     polaris.event-listener.types is prioritized
    */
   @Deprecated(since = "1.5.0", forRemoval = true)
   Optional<String> type();
 
   /**
-   * Comma separated list of event listers, each item must be a registered {@link
+   * Comma separated list of event listeners, each item must be a registered {@link
    * PolarisEventListener} identifier.
    */
   Optional<Set<String>> types();
+
+  /** Configuration of each event listener type. */
+  @WithParentName
+  Map<String, ListenerConfiguration> listenerConfig();
+
+  interface ListenerConfiguration {
+    /**
+     * Comma separated list of enabled event types. This event listener will only receive events of
+     * the selected types. If both the event types and event category configs are set, the listener
+     * will listen to both. If no listener configuration is present, then all event types are
+     * enabled.
+     */
+    Optional<Set<PolarisEventType>> enabledEventTypes();
+
+    /**
+     * Comma separated list of enabled event type categories. Each category is a collection of
+     * related Polaris event types. This event listener will only receive events of the selected
+     * event category, for example, consume only catalog events. If both the event types and event
+     * category configs are set, the listener will listen to both. If no listener configuration is
+     * present, then all event types are enabled.
+     */
+    Optional<Set<PolarisEventType.Category>> enabledEventCategories();
+  }
 }